Resolve rebase validation fallout
Preserve the rebased workflow checkpoint storage model under the split
model layout and keep SQLAlchemy runtime annotations explicit.

Split orchestration DTO and test modules that exceeded the branch lint
limits after incorporating `origin/main`, and raise the pytest timeout so
full-suite database fixture startup is not killed under shared host load.
