+++ tests/test_generation_orchestration_langgraph.py
@@ -21,9 +21,6 @@ from episodic.orchestration import (
 )
 from episodic.orchestration.langgraph import (
     GenerationGraphState,
-    _execute_node,
-    _finish_node,
-    _plan_node,
     build_generation_orchestration_graph,
     resume_generation_orchestration,
 )
@@ -203,7 +200,7 @@ class TestGenerationOrchestrationGraph:
         assert checkpoint is not None
         assert checkpoint.idempotency_key == suspended_result.idempotency_key
         assert checkpoint.payload["request"] == {"correlation_id": "corr-graph"}
-        assert tool_executor.calls == []
+        assert not tool_executor.calls
 
     @pytest.mark.asyncio
     async def test_generation_graph_reuses_checkpoint_for_same_step_key(self) -> None:
@@ -379,53 +376,3 @@ class TestGenerationOrchestrationGraph:
 
         with pytest.raises(ValueError, match="unknown checkpoint"):
             await store.mark_resumed(str(uuid.uuid4()))
-
-
-class TestLangGraphNodeValidation:
-    """Tests for individual LangGraph node state validation."""
-
-    @pytest.mark.asyncio
-    async def test_plan_node_requires_request(self) -> None:
-        """Planning node should fail loudly when request state is missing."""
-        with pytest.raises(ValueError, match="missing required state value: request"):
-            await _plan_node(
-                GenerationGraphState(),
-                planner=_FakePlanner(_planner_result()),
-            )
-
-    @pytest.mark.asyncio
-    async def test_execute_node_requires_request(self) -> None:
-        """Execution node should fail loudly when request state is missing."""
-        with pytest.raises(ValueError, match="missing required state value: request"):
-            await _execute_node(
-                GenerationGraphState(planner_result=_planner_result()),
-                tool_executor=_FakeToolExecutor(_action_result()),
-            )
-
-    @pytest.mark.asyncio
-    async def test_execute_node_requires_planner_result(self) -> None:
-        """Execution node should fail loudly when planning state is missing."""
-        with pytest.raises(
-            ValueError, match="missing required state value: planner_result"
-        ):
-            await _execute_node(
-                GenerationGraphState(request=_request()),
-                tool_executor=_FakeToolExecutor(_action_result()),
-            )
-
-    def test_finish_node_requires_request(self) -> None:
-        """Finish node should fail loudly when request state is missing."""
-        with pytest.raises(ValueError, match="missing required state value: request"):
-            _finish_node(
-                GenerationGraphState(
-                    planner_result=_planner_result(),
-                    action_results=(_action_result(),),
-                ),
-            )
-
-    def test_finish_node_requires_planner_result(self) -> None:
-        """Finish node should fail loudly when planning state is missing."""
-        with pytest.raises(
-            ValueError, match="missing required state value: planner_result"
-        ):
-            _finish_node(GenerationGraphState(request=_request()))
