+++ tests/test_orchestration_langgraph_properties.py
@@ -1,0 +1,102 @@
+"""Hypothesis property tests for orchestration LangGraph token rollups."""
+
+import string
+
+import hypothesis.strategies as st
+import pytest
+from hypothesis import given, settings
+
+from episodic.llm import LLMUsage
+from episodic.orchestration import (
+    ActionExecutionResult,
+    ActionKind,
+    ExecutionPlan,
+    GenerationGraphState,
+    GenerationOrchestrationRequest,
+    ModelTier,
+    PlannedAction,
+    PlannerResult,
+    build_generation_orchestration_graph,
+)
+from tests.test_orchestration_properties import (
+    _PropGraphPlanner,
+    _PropGraphToolExecutor,
+    _PropTokenInputs,
+    _token_inputs_strategy,
+)
+
+
+@given(
+    tokens=_token_inputs_strategy,
+    correlation_id=st.text(
+        min_size=1,
+        max_size=48,
+        alphabet=string.ascii_letters + string.digits + "-",
+    ),
+)
+@settings(max_examples=50)
+@pytest.mark.asyncio
+async def test_langgraph_total_tokens_non_negative(
+    tokens: _PropTokenInputs,
+    correlation_id: str,
+) -> None:
+    """Property test: LangGraph rollups keep total token counts semiring-safe."""
+    planner_usage = LLMUsage(
+        tokens.planner_input,
+        tokens.planner_output,
+        tokens.planner_input + tokens.planner_output,
+    )
+    tool_usage = LLMUsage(
+        tokens.action_input,
+        tokens.action_output,
+        tokens.action_input + tokens.action_output,
+    )
+    planner_result = PlannerResult(
+        plan=ExecutionPlan(
+            plan_version="1.0",
+            selected_planning_model="prop-plan-model",
+            selected_execution_model="prop-exec-model",
+            steps=(
+                PlannedAction(
+                    action_id="action-1",
+                    action_kind=ActionKind.GENERATE_SHOW_NOTES,
+                    rationale="prop graph rationale",
+                    model_tier=ModelTier.EXECUTION,
+                    required_inputs=("script_tei_xml",),
+                ),
+            ),
+        ),
+        usage=planner_usage,
+        model="gpt-4.1",
+        provider_response_id="prop-planner",
+        finish_reason="stop",
+    )
+    tool_result = ActionExecutionResult(
+        action_id="action-1",
+        action_kind=ActionKind.GENERATE_SHOW_NOTES,
+        model_tier=ModelTier.EXECUTION,
+        model="prop-exec-model",
+        summary="prop graph synthesis",
+        usage=tool_usage,
+    )
+
+    graph = build_generation_orchestration_graph(
+        planner=_PropGraphPlanner(result=planner_result),
+        tool_executor=_PropGraphToolExecutor(result=tool_result),
+    )
+
+    request = GenerationOrchestrationRequest(
+        correlation_id=correlation_id,
+        script_tei_xml=(
+            "<TEI><text><body><p>Hypothesis-driven graph workload</p></body>"
+            "</text></TEI>"
+        ),
+        template_structure=None,
+    )
+    state = await graph.ainvoke(GenerationGraphState(request=request))
+    orchestration_result = state["orchestration_result"]
+    expected_total_tokens = planner_usage.total_tokens + tool_usage.total_tokens
+    assert orchestration_result.total_usage.total_tokens >= 0
+    assert orchestration_result.total_usage.total_tokens == expected_total_tokens
+    assert state["planner_result"] == planner_result
+    assert state["action_results"][0].model == "prop-exec-model"
