+++ tests/test_orchestration_properties.py
@@ -17,7 +17,6 @@ from episodic.orchestration import (
     ActionExecutionResult,
     ActionKind,
     ExecutionPlan,
-    GenerationGraphState,
     GenerationOrchestrationConfig,
     GenerationOrchestrationRequest,
     ModelTier,
@@ -28,7 +27,6 @@ from episodic.orchestration import (
     StructuredGenerationPlanner,
     UnsupportedActionError,
     WorkflowStepIdentity,
-    build_generation_orchestration_graph,
     build_workflow_step_idempotency_key,
 )
 from episodic.orchestration.langgraph import (
@@ -345,79 +343,3 @@ async def test_planning_response_format_error_for_arbitrary_non_object_json(
         match=r"planner response must be an object",
     ):
         await planner.plan(request)
-
-
-@given(
-    tokens=_token_inputs_strategy,
-    correlation_id=st.text(
-        min_size=1,
-        max_size=48,
-        alphabet=string.ascii_letters + string.digits + "-",
-    ),
-)
-@settings(max_examples=50)
-@pytest.mark.asyncio
-async def test_langgraph_total_tokens_non_negative(
-    tokens: _PropTokenInputs,
-    correlation_id: str,
-) -> None:
-    """Property test: LangGraph rollups keep total token counts semiring-safe."""
-    planner_usage = LLMUsage(
-        tokens.planner_input,
-        tokens.planner_output,
-        tokens.planner_input + tokens.planner_output,
-    )
-    tool_usage = LLMUsage(
-        tokens.action_input,
-        tokens.action_output,
-        tokens.action_input + tokens.action_output,
-    )
-    planner_result = PlannerResult(
-        plan=ExecutionPlan(
-            plan_version="1.0",
-            selected_planning_model="prop-plan-model",
-            selected_execution_model="prop-exec-model",
-            steps=(
-                PlannedAction(
-                    action_id="action-1",
-                    action_kind=ActionKind.GENERATE_SHOW_NOTES,
-                    rationale="prop graph rationale",
-                    model_tier=ModelTier.EXECUTION,
-                    required_inputs=("script_tei_xml",),
-                ),
-            ),
-        ),
-        usage=planner_usage,
-        model="gpt-4.1",
-        provider_response_id="prop-planner",
-        finish_reason="stop",
-    )
-    tool_result = ActionExecutionResult(
-        action_id="action-1",
-        action_kind=ActionKind.GENERATE_SHOW_NOTES,
-        model_tier=ModelTier.EXECUTION,
-        model="prop-exec-model",
-        summary="prop graph synthesis",
-        usage=tool_usage,
-    )
-
-    graph = build_generation_orchestration_graph(
-        planner=_PropGraphPlanner(result=planner_result),
-        tool_executor=_PropGraphToolExecutor(result=tool_result),
-    )
-
-    request = GenerationOrchestrationRequest(
-        correlation_id=correlation_id,
-        script_tei_xml=(
-            "<TEI><text><body><p>Hypothesis-driven graph workload</p></body>"
-            "</text></TEI>"
-        ),
-        template_structure=None,
-    )
-    state = await graph.ainvoke(GenerationGraphState(request=request))
-    orchestration_result = state["orchestration_result"]
-    expected_total_tokens = planner_usage.total_tokens + tool_usage.total_tokens
-    assert orchestration_result.total_usage.total_tokens >= 0
-    assert orchestration_result.total_usage.total_tokens == expected_total_tokens
-    assert state["planner_result"] == planner_result
-    assert state["action_results"][0].model == "prop-exec-model"
